+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Ensure timeout is always reset on the context in the AfterSuite to ensure enough time is given for cleanup
+
 ### Changed
 
 - Updated teleport api module to latest available.