+++ CHANGELOG.md
@@ -22,6 +22,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Remove node checks from CAPA upgrade suite because we already check nodes in common tests after the upgrade.
 
+### Removed
+
+- Remove node checks from upgrade because we already check nodes in common tests after the upgrade.
+
 ## [1.47.0] - 2024-06-14
 
 ### Added